Handle error from Connect() to IRC server
Fix hanging on failed connect.

In main(), if c.Connect() fails, the program will proceed to the
for-select loop, but neither of the channels listened on will fire. Even
if the `quit` channel does fire (SIGINT/SIGTERM), it won't shut down
the program but just send the QUIT command on a non-existent connection
without returning an error.

To fix this:
 - return immediately on failed connect
 - only use Quit() if connection to server is active, otherwise return

The second part is vulnerable to race conditions between Connected() and
Quit().